Update nxos.py
Added in support for Get Fex Status
This will allow for a user to review the fex's and what the status or health of the ports are for that fex.  

Improvements:
Got it working, but the loop could be better based on the number of times "fex_number" shows up to get an accurate count.  Right now it will loop until it errors, and that is the expected behavior of it being complete. 

Example:
[root@vm12 Napalm]# python napalm_nxos_get_fex_status.py
[{'fex_number': '122', 'fex_state': ['Po122', 'Eth1/11', 'Eth1/35', 'Active', 'Active', 'Active']}