Update nxos.py
Added in fuction get_fex_information

This will give the user more information on fex's ( number, description, state,model, and serial number of fex connected devices )

Improvements:
The loop was changed for this and the other get_fex_* to be 99, while the fex is 100-199, it's not 200.  so the max the number of fex's possible per device is 100.  math is hard.
The loop is horrid as the expected result is for it to fail, has room for improvement, was needed for incrementing on i. but it works

Test Results:
[root@vm12 Napalm]# python napalm_nxos_get_fex.py
[{'fex_number': '122', 'fex_serial': 'FOX19XXXXX', 'fex_model': 'N2K-C2248TP-E-1GE', 'fex_state': 'Online', 'fex_description': 'ATL-LABLAB1'},